+++ src/game/MotionMaster.h
@@ -89,7 +89,7 @@ class TRINITY_DLL_SPEC MotionMaster //: private std::stack<MovementGenerator *>
 
         explicit MotionMaster(Unit *unit) : i_owner(unit), m_expList(NULL), m_cleanFlag(MMCF_NONE), i_top(-1)
         {
-            for(int i = 0; i < MAX_MOTION_SLOT; ++i)
+            for(uint8 i = 0; i < MAX_MOTION_SLOT; ++i)
             {
                 Impl[i] = NULL;
                 needInit[i] = true;
